% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion grafica mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Creación de gráfico de temperatura promedio mensual por estación}
\usage{
grafico_temperatura_mensual(datos, colores = NULL, titulo = "Temperatura")
}
\arguments{
\item{datos}{data.frame que contiene los datos. Debe incluir las columnas:
\code{estacion}: Un factor o cadena que indica la estación del año.
\code{mes}: mes del año de forma numerica.
\code{temperatura_abrigo_150cm}: Un valor numérico que indica la temperatura medida.}

\item{colores}{en el caso de querer presonalizar los colores del grafico.}

\item{titulo}{Un string que especifica el título del gráfico.}
}
\value{
Un grafico de líneas con la temperatura promedio mensual para cada estación.
}
\description{
La función \code{grafica_temperatura_mensual} genera un gráfico de líneas que muestra la temperatura promedio mensual agrupada por estación.
}
\details{
Se puede personalizar el título del gráfico y los colores de las líneas.
}
\examples{
datos_clima <- data.frame(
  estacion = rep(c("Verano", "Otoño", "Invierno", "Primavera"), each = 12),
  mes = rep(1:12, times = 4),
  temperatura_abrigo_150cm = runif(48, min = -5, max = 35)
)

ejemplo_1
 grafico_temperatura_mensual(datos_clima)

ejemplo 2

grafico_temperatura_mensual(datos_clima, colores = c("red", "blue", "green", "purple"), titulo = "Temperatura")


}
