% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadatos_completos}
\alias{metadatos_completos}
\title{Metadatos Completos Dataset}
\format{
Un data frame con las siguientes columnas:
\describe{
\item{ID}{Identificador del dataset.}
\item{Variable}{Nombre de la variable en el dataset.}
\item{Descripcion}{Descripción de la variable.}
\item{Unidad}{Unidad de medida de la variable.}
}
}
\source{
https://raw.githubusercontent.com/rse-r/intro-programacion/main/datos/metadatos_completos.csv
}
\usage{
metadatos_completos
}
\description{
Descripción del dataset de metadatos completos.
}
\keyword{datasets}
