% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NH0046}
\alias{NH0046}
\title{NH0046 Dataset}
\format{
Un data frame con las siguientes columnas:
\describe{
\item{Fecha}{Fecha de la observación.}
\item{Temperatura}{Temperatura registrada (en grados Celsius).}
\item{Humedad}{Porcentaje de humedad relativa.}
\item{Presion}{Presión atmosférica (en hPa).}
}
}
\source{
https://raw.githubusercontent.com/rse-r/intro-programacion/main/datos/NH0046.csv
}
\usage{
NH0046
}
\description{
Descripción del dataset NH0046.
}
\keyword{datasets}
