% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NH0098}
\alias{NH0098}
\title{NH0098 Dataset}
\format{
Un data frame con las siguientes columnas:
\describe{
\item{Fecha}{Fecha de la observación.}
\item{Temperatura}{Temperatura registrada (en grados Celsius).}
\item{Punto_Rocio}{Punto de rocío (en grados Celsius).}
\item{Viento}{Velocidad del viento (en km/h).}
}
}
\source{
https://raw.githubusercontent.com/rse-r/intro-programacion/main/datos/NH0098.csv
}
\usage{
NH0098
}
\description{
Descripción del dataset NH0098.
}
\keyword{datasets}
