% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NH0437}
\alias{NH0437}
\title{NH0437 Dataset}
\format{
Un data frame con las siguientes columnas:
\describe{
\item{Fecha}{Fecha de la observación.}
\item{Temperatura_Maxima}{Temperatura máxima registrada (en grados Celsius).}
\item{Temperatura_Minima}{Temperatura mínima registrada (en grados Celsius).}
\item{Radiacion}{Radiación solar (en W/m²).}
}
}
\source{
https://raw.githubusercontent.com/rse-r/intro-programacion/main/datos/NH0437.csv
}
\usage{
NH0437
}
\description{
Descripción del dataset NH0437.
}
\keyword{datasets}
