% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NH0472}
\alias{NH0472}
\title{NH0472 Dataset}
\format{
Un data frame con las siguientes columnas:
\describe{
\item{Fecha}{Fecha de la observación.}
\item{Temperatura}{Temperatura registrada (en grados Celsius).}
\item{Humedad}{Porcentaje de humedad relativa.}
\item{Precipitacion}{Precipitación (en mm).}
}
}
\source{
https://raw.githubusercontent.com/rse-r/intro-programacion/main/datos/NH0472.csv
}
\usage{
NH0472
}
\description{
Descripción del dataset NH0472.
}
\keyword{datasets}
