% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NH0910}
\alias{NH0910}
\title{Descripción del dataset NH0910.}
\format{
Un data frame con las siguientes columnas:
\describe{
\item{Fecha}{Fecha de la observación.}
\item{Temperatura}{Temperatura registrada (en grados Celsius).}
\item{Humedad}{Porcentaje de humedad relativa.}
\item{Viento}{Velocidad del viento (en km/h).}
\item{Presion}{Presión atmosférica (en hPa).}
}
}
\source{
https://raw.githubusercontent.com/rse-r/intro-programacion/main/datos/NH0910.csv
}
\usage{
NH0910
}
\description{
Descripción del dataset NH0910.
}
\keyword{datasets}
