% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion tabla de resumen de temperatura.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Creación de tabla que resume temperatura de abrigo}
\usage{
tabla_resumen_temperatura(datos)
}
\arguments{
\item{datos}{data.frame de las estaciones en las que queremos resumir sus temperatura abrigo y hacer la tabla resumen}
}
\value{
Devuelve una tabla en forma de resumen de la temperatura abrigo de las estaciones ingresadas
}
\description{
La función \code{tabla_resumen_temperatura} crea una tabla que resume la temperatura_abrigo_150cm para las diferentes estaciones
}
\examples{
tabla_resumen_temperatura(datos)

}
