% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion convertidora de f a c.R
\docType{data}
\name{temp_max_hoy}
\alias{temp_max_hoy}
\title{Convierte temperatura de farenheit a centigrados}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
temp_max_hoy
}
\arguments{
\item{temp_max_hoy}{es el valor a convertir}
}
\value{
Un vector numérico que es el valor de la temperatura en grados celcius
}
\description{
La función \code{convertir_farenheit_a_centigrados} toma una temperatura establecida en farenheit y la convierte a grados celcius.
}
\examples{
convertir_farenheit_a_centigrados(90)
convertir_farenheit_a_centigrados(75)

}
\keyword{datasets}
