% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos.R
\name{descargar_datos}
\alias{descargar_datos}
\title{Descargar datos meteorológicos}
\usage{
descargar_datos(id_estacion, directorio_destino)
}
\arguments{
\item{id_estacion}{ID de la estación.}

\item{directorio_destino}{Directorio donde se guardará el archivo CSV.}
}
\value{
Un archivo CSV con los datos meteorológicos de la estación con el ID ingresado.
}
\description{
La función \code{descargar_datos} toma el ID de una estación y un directorio de destino, y descarga en dicha ubicación un archivo CSV con los datos meteorológicos de la estación correspondiente. Los IDs posibles son: NH0098, NH0046, NH437, NH472 y NH0910.
}
\examples{
descargar_datos("NH0098", "data")
descargar_datos("NH0910", "data")
}
