% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion_convertidora_de_f_a_c.R
\name{farenheit_a_centigrados}
\alias{farenheit_a_centigrados}
\title{Convierte temperatura de farenheit a centigrados}
\usage{
farenheit_a_centigrados(temperatura_fahrenheit)
}
\arguments{
\item{temperatura_fahrenheit}{numerico. Valor de temperatura en grados Fahrenheit que se desea convertir a grados Celsius.
Debe ser un valor numérico; de lo contrario, la función devolverá un mensaje de error.}
}
\value{
Un vector numérico que es el valor de la temperatura en grados celcius
Si el valor de \code{temperatura_fahrenheit} no es numérico, la función lanza un error detallado.
}
\description{
La función \code{farenheit_a_centigrados} convierte un valor de temperatura de grados Fahrenheit a grados Celsius.
}
\examples{
#convertir 90 grados fahrenheit a celsius
farenheit_a_centigrados(90)
#convertir 75 grados fahrenheit a celsius
farenheit_a_centigrados(75)

}
