% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion_grafica_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Creación de gráfico de temperatura promedio mensual por estación}
\usage{
grafico_temperatura_mensual(datos, colores = NULL, titulo = "Temperatura")
}
\arguments{
\item{datos}{data.frame que contiene los datos. Debe incluir las columnas:
\code{estacion}: Un factor o cadena que indica la estación del año.
\code{mes}: mes del año de forma numerica (id).
\code{temperatura_abrigo_150cm}: Un valor numérico que indica la temperatura medida.}

\item{colores}{en el caso de querer presonalizar los colores del grafico.}

\item{titulo}{Un string que especifica el título del gráfico.}
}
\value{
Un grafico de líneas con la temperatura promedio mensual para cada estación.
}
\description{
La función \code{grafica_temperatura_mensual} genera un gráfico de líneas que muestra la temperatura promedio mensual agrupada por estación.
}
\details{
Se puede personalizar el título del gráfico y los colores de las líneas.
}
\examples{

grafico_temperatura_mensual(NH0910, "pink", "Temperatura")


}
