% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion_tabla_de_resumen_de_temperatura.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Creación de tabla que resume temperatura de abrigo}
\usage{
tabla_resumen_temperatura(datos)
}
\arguments{
\item{datos}{data.frame de las estaciones en las que queremos resumir sus temperatura abrigo y hacer la tabla resumen
Es obligatorio que sea un data frame, de lo contrario, la función lanzará un mensaje de error.}
}
\value{
Una tabla con el resumen estadístico por estación, incluyendo las siguientes columnas:
\itemize{
\item \code{estacion}: Identificador de la estación.
\item \code{min_temp}: Temperatura mínima registrada para la estación (considerando NA en los datos).
\item \code{max_temp}: Temperatura máxima registrada para la estación.
\item \code{mean_temp}: Temperatura media registrada para la estación.
}
}
\description{
La función \code{tabla_resumen_temperatura}  calcula un resumen estadístico (mínimo, máximo y promedio) de la
temperatura registrada a 150 cm de altura para cada estación.
}
\examples{
tabla_resumen_temperatura(NH0910)

}
